housekeeping in fileio. Moved some functions to parent class to avoid repeating. Defined abstract functions in parent class. this helps when building a new child class.
