+++ mdsuite/file_io/trajectory_files.py
@@ -125,3 +125,44 @@ class TrajectoryFile(FileProcessor, metaclass=abc.ABCMeta):
         #     np.linspace(0, self.project.number_of_atoms, dtype - int)
 
         return structure
+
+    @staticmethod
+    def _build_architecture(species_summary: dict, property_groups: dict, number_of_configurations: int):
+        """
+        Build the database_path architecture for use by the database_path class
+
+        Parameters
+        ----------
+        species_summary : dict
+                Species summary passed to the experiment class
+        property_groups : dict
+                Property information passed to the experiment class
+        number_of_configurations : int
+                Number of configurations in the file
+
+        """
+        architecture = {}  # instantiate the database_path architecture dictionary
+        for species in species_summary:
+            architecture[species] = {}
+            for observable in property_groups:
+                architecture[species][observable] = (len(species_summary[species]['indices']),
+                                                     number_of_configurations,
+                                                     len(property_groups[observable]))
+
+        return architecture
+
+    @abc.abstractmethod
+    def _get_species_information(self):
+        pass
+
+    @abc.abstractmethod
+    def _get_time_information(self):
+        pass
+
+    @abc.abstractmethod
+    def _get_number_of_configurations(self):
+        pass
+
+    @abc.abstractmethod
+    def _get_number_of_atoms(self):
+        pass
